WFS provider: declare its QgsSQLComposerDialog implementation as a subset string editor provider
