+++ src/providers/wfs/qgswfsprovidergui.cpp
@@ -18,7 +18,8 @@
 #include "qgswfssourceselect.h"
 #include "qgssourceselectprovider.h"
 #include "qgsproviderguimetadata.h"
-
+#include "qgssubsetstringeditorprovider.h"
+#include "qgswfssubsetstringeditor.h"
 
 //! Provider for WFS layers source select
 class QgsWfsSourceSelectProvider : public QgsSourceSelectProvider
@@ -32,6 +33,29 @@ class QgsWfsSourceSelectProvider : public QgsSourceSelectProvider
     QgsAbstractDataSourceWidget *createDataSourceWidget( QWidget *parent = nullptr, Qt::WindowFlags fl = Qt::Widget, QgsProviderRegistry::WidgetMode widgetMode = QgsProviderRegistry::WidgetMode::Embedded ) const override
     {
       return new QgsWFSSourceSelect( parent, fl, widgetMode );
+    }
+};
+
+//! Provider for dedicated subset string editor for WFS layers
+class QgsWfsSubsetStringEditorProvider: public QgsSubsetStringEditorProvider
+{
+  public:
+
+    QString providerKey() const override { return QgsWFSProvider::WFS_PROVIDER_KEY; }
+
+    bool canHandleLayer( QgsVectorLayer *layer ) const override
+    {
+      QgsDataProvider *provider = layer->dataProvider();
+      return static_cast< bool >( dynamic_cast<QgsWFSProvider *>( provider ) );
+    }
+
+    QgsSubsetStringEditorInterface *createDialog( QgsVectorLayer *layer, QWidget *parent, Qt::WindowFlags fl ) override
+    {
+      QgsDataProvider *provider = layer->dataProvider();
+      QgsWFSProvider *wfsProvider = dynamic_cast<QgsWFSProvider *>( provider );
+      if ( !wfsProvider )
+        return nullptr;
+      return QgsWfsSubsetStringEditor::create( layer, wfsProvider, parent, fl );
     }
 };
 
@@ -55,6 +79,12 @@ class QgsWfsProviderGuiMetadata: public QgsProviderGuiMetadata
     {
       return QList<QgsDataItemGuiProvider *>()
              << new QgsWfsDataItemGuiProvider;
+    }
+
+    QList<QgsSubsetStringEditorProvider *> subsetStringEditorProviders() override
+    {
+      return QList<QgsSubsetStringEditorProvider *>()
+             << new QgsWfsSubsetStringEditorProvider;
     }
 
 };
