+++ src/providers/wfs/qgswfssourceselect.cpp
@@ -35,6 +35,7 @@
 #include "qgsgui.h"
 #include "qgsquerybuilder.h"
 #include "qgswfsguiutils.h"
+#include "qgswfssubsetstringeditor.h"
 
 #include <QDomDocument>
 #include <QListWidgetItem>
@@ -531,82 +532,6 @@ void QgsWFSSourceSelect::addButtonClicked()
   }
 }
 
-QgsWFSValidatorCallback::QgsWFSValidatorCallback( QObject *parent,
-    const QgsWFSDataSourceURI &uri,
-    const QString &allSql,
-    const QgsWfsCapabilities::Capabilities &caps )
-  : QObject( parent )
-  , mURI( uri )
-  , mAllSql( allSql )
-  , mCaps( caps )
-{
-}
-
-bool QgsWFSValidatorCallback::isValid( const QString &sqlStr, QString &errorReason, QString &warningMsg )
-{
-  errorReason.clear();
-  if ( sqlStr.isEmpty() || sqlStr == mAllSql )
-    return true;
-
-  QgsWFSDataSourceURI uri( mURI );
-  uri.setSql( sqlStr );
-
-  QgsDataProvider::ProviderOptions options;
-  QgsWFSProvider p( uri.uri(), options, mCaps );
-  if ( !p.isValid() )
-  {
-    errorReason = p.processSQLErrorMsg();
-    return false;
-  }
-  warningMsg = p.processSQLWarningMsg();
-
-  return true;
-}
-
-QgsWFSTableSelectedCallback::QgsWFSTableSelectedCallback( QgsSQLComposerDialog *dialog,
-    const QgsWFSDataSourceURI &uri,
-    const QgsWfsCapabilities::Capabilities &caps )
-  : QObject( dialog )
-  , mDialog( dialog )
-  , mURI( uri )
-  , mCaps( caps )
-{
-}
-
-void QgsWFSTableSelectedCallback::tableSelected( const QString &name )
-{
-  QString typeName( QgsSQLStatement::stripQuotedIdentifier( name ) );
-  QString prefixedTypename( mCaps.addPrefixIfNeeded( typeName ) );
-  if ( prefixedTypename.isEmpty() )
-    return;
-  QgsWFSDataSourceURI uri( mURI );
-  uri.setTypeName( prefixedTypename );
-
-  QgsDataProvider::ProviderOptions providerOptions;
-  QgsWFSProvider p( uri.uri(), providerOptions, mCaps );
-  if ( !p.isValid() )
-  {
-    return;
-  }
-
-  QList< QgsSQLComposerDialog::PairNameType> fieldList;
-  QString fieldNamePrefix( QgsSQLStatement::quotedIdentifierIfNeeded( typeName ) + "." );
-  const auto constToList = p.fields().toList();
-  for ( const QgsField &field : constToList )
-  {
-    QString fieldName( fieldNamePrefix + QgsSQLStatement::quotedIdentifierIfNeeded( field.name() ) );
-    fieldList << QgsSQLComposerDialog::PairNameType( fieldName, field.typeName() );
-  }
-  if ( !p.geometryAttribute().isEmpty() )
-  {
-    QString fieldName( fieldNamePrefix + QgsSQLStatement::quotedIdentifierIfNeeded( p.geometryAttribute() ) );
-    fieldList << QgsSQLComposerDialog::PairNameType( fieldName, QStringLiteral( "geometry" ) );
-  }
-  fieldList << QgsSQLComposerDialog::PairNameType( fieldNamePrefix + "*", QString() );
-
-  mDialog->addColumnNames( fieldList, name );
-}
-
 void QgsWFSSourceSelect::buildQuery( const QModelIndex &index )
 {
   if ( !index.isValid() )
@@ -681,97 +606,9 @@ void QgsWFSSourceSelect::buildQuery( const QModelIndex &index )
     sql = allSql;
   }
 
-  QgsSQLComposerDialog *d = new QgsSQLComposerDialog( this );
+  auto d = QgsWfsSubsetStringEditor::create( nullptr, &p, this );
 
-  QgsWFSValidatorCallback *validatorCbk = new QgsWFSValidatorCallback( d, uri, allSql, mCaps );
-  d->setSQLValidatorCallback( validatorCbk );
-
-  QgsWFSTableSelectedCallback *tableSelectedCbk = new QgsWFSTableSelectedCallback( d, uri, mCaps );
-  d->setTableSelectedCallback( tableSelectedCbk );
-
-  const bool bSupportJoins = mCaps.featureTypes.size() > 1 && mCaps.supportsJoins;
-  d->setSupportMultipleTables( bSupportJoins, QgsSQLStatement::quotedIdentifierIfNeeded( displayedTypeName ) );
-
-  QMap< QString, QString > mapTypenameToTitle;
-  Q_FOREACH ( const QgsWfsCapabilities::FeatureType f, mCaps.featureTypes )
-    mapTypenameToTitle[f.name] = f.title;
-
-  QList< QgsSQLComposerDialog::PairNameTitle > tablenames;
-  tablenames << QgsSQLComposerDialog::PairNameTitle(
-               QgsSQLStatement::quotedIdentifierIfNeeded( displayedTypeName ), mapTypenameToTitle[typeName] );
-  if ( bSupportJoins )
-  {
-    for ( int i = 0; i < mModel->rowCount(); i++ )
-    {
-      const QString iterTypename = mModel->index( i, MODEL_IDX_NAME ).data().toString();
-      if ( iterTypename != typeName )
-      {
-        QString displayedIterTypename( iterTypename );
-        QString unprefixedIterTypename( QgsWFSUtils::removeNamespacePrefix( iterTypename ) );
-        if ( !mCaps.setAmbiguousUnprefixedTypename.contains( unprefixedIterTypename ) )
-          displayedIterTypename = unprefixedIterTypename;
-
-        tablenames << QgsSQLComposerDialog::PairNameTitle(
-                     QgsSQLStatement::quotedIdentifierIfNeeded( displayedIterTypename ), mapTypenameToTitle[iterTypename] );
-      }
-    }
-  }
-  d->addTableNames( tablenames );
-
-  QList< QgsSQLComposerDialog::Function> functionList;
-  Q_FOREACH ( const QgsWfsCapabilities::Function &f, mCaps.functionList )
-  {
-    QgsSQLComposerDialog::Function dialogF;
-    dialogF.name = f.name;
-    dialogF.returnType = f.returnType;
-    dialogF.minArgs = f.minArgs;
-    dialogF.maxArgs = f.maxArgs;
-    Q_FOREACH ( const QgsWfsCapabilities::Argument &arg, f.argumentList )
-    {
-      dialogF.argumentList << QgsSQLComposerDialog::Argument( arg.name, arg.type );
-    }
-    functionList << dialogF;
-  }
-  d->addFunctions( functionList );
-
-  QList< QgsSQLComposerDialog::Function> spatialPredicateList;
-  Q_FOREACH ( const QgsWfsCapabilities::Function &f, mCaps.spatialPredicatesList )
-  {
-    QgsSQLComposerDialog::Function dialogF;
-    dialogF.name = f.name;
-    dialogF.returnType = f.returnType;
-    dialogF.minArgs = f.minArgs;
-    dialogF.maxArgs = f.maxArgs;
-    Q_FOREACH ( const QgsWfsCapabilities::Argument &arg, f.argumentList )
-    {
-      dialogF.argumentList << QgsSQLComposerDialog::Argument( arg.name, arg.type );
-    }
-    spatialPredicateList << dialogF;
-  }
-  d->addSpatialPredicates( spatialPredicateList );
-
-  QList< QgsSQLComposerDialog::PairNameType> fieldList;
-  QString fieldNamePrefix;
-  if ( bSupportJoins )
-  {
-    fieldNamePrefix = QgsSQLStatement::quotedIdentifierIfNeeded( displayedTypeName ) + ".";
-  }
-  const auto constToList = p.fields().toList();
-  for ( const QgsField &field : constToList )
-  {
-    QString fieldName( fieldNamePrefix + QgsSQLStatement::quotedIdentifierIfNeeded( field.name() ) );
-    fieldList << QgsSQLComposerDialog::PairNameType( fieldName, field.typeName() );
-  }
-  if ( !p.geometryAttribute().isEmpty() )
-  {
-    QString fieldName( fieldNamePrefix + QgsSQLStatement::quotedIdentifierIfNeeded( p.geometryAttribute() ) );
-    fieldList << QgsSQLComposerDialog::PairNameType( fieldName, QStringLiteral( "geometry" ) );
-  }
-  fieldList << QgsSQLComposerDialog::PairNameType( fieldNamePrefix + "*", QString() );
-
-  d->addColumnNames( fieldList, QgsSQLStatement::quotedIdentifierIfNeeded( displayedTypeName ) );
-
-  d->setSql( sql );
+  d->setSubsetString( sql );
 
   mSQLIndex = index;
   mSQLComposerDialog = d;
@@ -803,7 +640,7 @@ void QgsWFSSourceSelect::updateSql()
   const QString typeName = mSQLIndex.sibling( mSQLIndex.row(), MODEL_IDX_NAME ).data().toString();
   QModelIndex filterIndex = mSQLIndex.sibling( mSQLIndex.row(), MODEL_IDX_SQL );
 
-  QString sql = mSQLComposerDialog->sql();
+  QString sql = mSQLComposerDialog->subsetString();
   mSQLComposerDialog = nullptr;
 
   QString displayedTypeName( typeName );
