+++ src/providers/wfs/qgswfssourceselect.h
@@ -25,7 +25,6 @@
 #include "qgsoapifcollection.h"
 #include "qgsproviderregistry.h"
 #include "qgsabstractdatasourcewidget.h"
-#include "qgssqlcomposerdialog.h"
 
 #include <QItemDelegate>
 #include <QStandardItemModel>
@@ -33,6 +32,7 @@
 
 class QgsProjectionSelectionDialog;
 class QgsWfsCapabilities;
+class QgsSubsetStringEditorInterface;
 
 class QgsWFSItemDelegate : public QItemDelegate
 {
@@ -43,21 +43,6 @@ class QgsWFSItemDelegate : public QItemDelegate
 
     QSize sizeHint( const QStyleOptionViewItem &option, const QModelIndex &index ) const override;
 
-};
-
-class QgsWFSValidatorCallback: public QObject, public QgsSQLComposerDialog::SQLValidatorCallback
-{
-    Q_OBJECT
-
-  public:
-    QgsWFSValidatorCallback( QObject *parent,
-                             const QgsWFSDataSourceURI &uri, const QString &allSql,
-                             const QgsWfsCapabilities::Capabilities &caps );
-    bool isValid( const QString &sql, QString &errorReason, QString &warningMsg ) override;
-  private:
-    QgsWFSDataSourceURI mURI;
-    QString mAllSql;
-    const QgsWfsCapabilities::Capabilities &mCaps;
 };
 
 class QgsWFSSourceSelect: public QgsAbstractDataSourceWidget, private Ui::QgsWFSSourceSelectBase
@@ -91,7 +76,7 @@ class QgsWFSSourceSelect: public QgsAbstractDataSourceWidget, private Ui::QgsWFS
     QPushButton *mBuildQueryButton = nullptr;
     QgsWfsCapabilities::Capabilities mCaps;
     QModelIndex mSQLIndex;
-    QgsSQLComposerDialog *mSQLComposerDialog = nullptr;
+    QgsSubsetStringEditorInterface *mSQLComposerDialog = nullptr;
     QString mVersion;
 
     /**
@@ -137,23 +122,6 @@ class QgsWFSSourceSelect: public QgsAbstractDataSourceWidget, private Ui::QgsWFS
     void startOapifCollectionsRequest( const QString &url );
     void resizeTreeViewAfterModelFill();
     bool isOapif() const { return mVersion == QLatin1String( "OGC_API_FEATURES" ); }
-};
-
-
-class QgsWFSTableSelectedCallback: public QObject, public QgsSQLComposerDialog::TableSelectedCallback
-{
-    Q_OBJECT
-
-  public:
-    QgsWFSTableSelectedCallback( QgsSQLComposerDialog *dialog,
-                                 const QgsWFSDataSourceURI &uri,
-                                 const QgsWfsCapabilities::Capabilities &caps );
-    void tableSelected( const QString &name ) override;
-
-  private:
-    QgsSQLComposerDialog *mDialog = nullptr;
-    QgsWFSDataSourceURI mURI;
-    const QgsWfsCapabilities::Capabilities &mCaps;
 };
 
 #endif
